requirements.txt
Django
MySQL-python
South
django-registration
django-email-as-username==1.4.6
mailsnake
django-tastypie